Use let() instead of let!()
This commit changes the sample gcloud integration test from using let!()
(which is evaluated immediately and allows for value changes between
tests) to let() (which is lazy-evaluted and memoized). Unless there's
a reason for allowing the value declared to be changed between tests
it's better to err on the size of memoizing.